+++ httpx_auth/__init__.py
@@ -15,6 +15,7 @@ from httpx_auth.authentication import (
     OAuth2ClientCredentials,
     OktaClientCredentials,
     OAuth2ResourceOwnerPasswordCredentials,
+    OktaResourceOwnerPasswordCredentials,
     WakaTimeAuthorizationCode,
 )
 from httpx_auth.oauth2_tokens import JsonTokenFileCache
@@ -47,6 +48,7 @@ __all__ = [
     "OAuth2ClientCredentials",
     "OktaClientCredentials",
     "OAuth2ResourceOwnerPasswordCredentials",
+    "OktaResourceOwnerPasswordCredentials",
     "WakaTimeAuthorizationCode",
     "JsonTokenFileCache",
     "AWS4Auth",